requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
babel==2.17.0
beautifulsoup4==4.12.2
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
courlan==1.3.2
cryptography==46.0.3
dateparser==1.2.2
ddgs==9.5.5
decorator==5.2.1
distro==1.9.0
dnslib==0.9.26
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.100.0
filelock==3.20.1
groq==1.0.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jusText==3.0.2
kombu==5.6.1
lxml==6.0.2
lxml_html_clean==0.4.3
msgpack==1.1.2
packaging==25.0
primp==0.15.0
prompt_toolkit==3.0.52
py==1.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymetasploit3==1.0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==7.1.0
regex==2025.11.3
requests==2.31.0
requests-file==3.0.1
retry==0.9.2
shodan==1.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
tld==0.13.1
tldextract==5.3.0
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.14
websockets==11.0.3
xlsxwriter==3.2.9
